+++ HISTORY.rst
@@ -6,6 +6,9 @@ History
 --------------
 * Refactor info.py into environment.py to account for regression in cpython micro release (#5) [PR #6]
 * Change file format from yaml to json with sorted keys (#4) [PR #7]
+* Remove pyenvdiff.com as default server. [PR #8]
+* Simplify user-created custom Collector. [PR #8]
+* Enable user-created custom POST command. [PR #8]
 
 0.1.0 (2016-12-24)
 ------------------